Implement q_insert_tail, q_reverse, q_size
In order to implement q_insert_tail and q_size in O(1), I add two
elements, *tail and size, to the queue struct.

q_insert_tail:
If q, newt, or value of newt is NULL, return NULL. Then assign the
newt to the tail of queue.

q_size:
Check if the queue is NULL, and then return q->size.

q_reverse:
Change the next of a node to the previous one. Exchange the tail and
the head of queue.