Implement q_remove_head
First check whether q is NULL and compute how many char I can store.
I use a tmp to store the head of the queue and assign the next one as
new head of the queue. Then use memcpy() function to copy tmp
value to sp, and assign last element of sp as '\0'.